added NSCustomClass to simplify the logic
