+++ Src/ImageQT/DllInterop/Mac/ObjectCRuntime.cs
@@ -21,7 +21,7 @@ internal partial class ObjectCRuntime
     public static partial void ObjCMsgSend(IntPtr receiver, IntPtr selector, IntPtr arg1);
 
     [LibraryImport(_dllName, EntryPoint = "objc_msgSend")]
-    public static partial void ObjCMsgSend(IntPtr receiver, IntPtr selector, CustomMethodDelegate arg1);
+    public static partial void ObjCMsgSend(IntPtr receiver, IntPtr selector, NSCustomClass arg1);
 
     [LibraryImport(_dllName, EntryPoint = "objc_msgSend")]
     public static partial void ObjCMsgSend(IntPtr receiver, IntPtr selector, [MarshalAs(UnmanagedType.Bool)] bool arg1);
